feat(ui/home): link home ui to view model (not working)
