+++ app/src/androidTest/java/com/android/unio/screen/EventListOverviewTest.kt
@@ -6,15 +6,19 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.unit.ExperimentalUnitApi
+import androidx.navigation.NavHostController
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.android.unio.model.event.Event
 import com.android.unio.model.event.EventListViewModel
 import com.android.unio.model.event.EventRepositoryMock
-import com.android.unio.ui.event.EventListOverview
+import com.android.unio.ui.home.HomeScreen
+import com.android.unio.ui.navigation.NavigationAction
 import kotlinx.coroutines.test.runBlockingTest
+import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mockito.kotlin.mock
 
 /**
  * Test class for the EventListOverview Composable. This class contains unit tests to validate the
@@ -22,12 +26,20 @@ import org.junit.runner.RunWith
  */
 @ExperimentalUnitApi
 @RunWith(AndroidJUnit4::class)
-class EventListOverviewTest {
+class HomeScreenTest {
 
   @get:Rule val composeTestRule = createComposeRule()
 
   // Mock event repository to provide test data.
   private val mockEventRepository = EventRepositoryMock()
+  private lateinit var navHostController: NavHostController
+  private lateinit var navigationAction: NavigationAction
+
+  @Before
+  fun setUp() {
+    navHostController = mock { NavHostController::class.java }
+    navigationAction = NavigationAction(navHostController)
+  }
 
   /**
    * Tests the functionality of switching between tabs and verifying animations. Ensures that the
@@ -38,7 +50,11 @@ class EventListOverviewTest {
   fun testTabSwitchingAndAnimation() {
     composeTestRule.setContent {
       val eventListViewModel = EventListViewModel(mockEventRepository)
-      EventListOverview(eventListViewModel = eventListViewModel, onAddEvent = {}, onEventClick = {})
+      HomeScreen(
+        navigationAction,
+        eventListViewModel = eventListViewModel,
+        onAddEvent = {},
+        onEventClick = {})
     }
 
     // Assert that the 'All' tab exists and has a click action.
@@ -74,7 +90,11 @@ class EventListOverviewTest {
             }
           }
       val eventListViewModel = EventListViewModel(emptyEventRepository)
-      EventListOverview(eventListViewModel = eventListViewModel, onAddEvent = {}, onEventClick = {})
+      HomeScreen(
+        navigationAction,
+        eventListViewModel = eventListViewModel,
+        onAddEvent = {},
+        onEventClick = {})
     }
 
     // Assert that the empty event prompt is displayed.
@@ -92,7 +112,8 @@ class EventListOverviewTest {
 
     composeTestRule.setContent {
       val eventListViewModel = EventListViewModel(mockEventRepository)
-      EventListOverview(
+      HomeScreen(
+        navigationAction = navigationAction,
           eventListViewModel = eventListViewModel,
           onAddEvent = { mapClicked = true },
           onEventClick = {})
@@ -113,7 +134,8 @@ class EventListOverviewTest {
 
     composeTestRule.setContent {
       val eventListViewModel = EventListViewModel(mockEventRepository)
-      EventListOverview(
+      HomeScreen(
+        navigationAction,
           eventListViewModel = eventListViewModel,
           onAddEvent = { addClicked = true },
           onEventClick = {})
