+++ app/src/main/java/com/android/unio/MainActivity.kt
@@ -14,6 +14,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navigation
+import com.android.unio.model.event.EventListViewModel
+import com.android.unio.model.event.EventRepositoryFirestore
 import com.android.unio.ui.association.AssociationProfile
 import com.android.unio.ui.authentication.AccountDetails
 import com.android.unio.ui.authentication.EmailVerificationScreen
@@ -28,6 +30,7 @@ import com.android.unio.ui.theme.AppTheme
 import com.android.unio.ui.user.UserProfileScreen
 import com.google.firebase.Firebase
 import com.google.firebase.auth.auth
+import com.google.firebase.firestore.FirebaseFirestore
 
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
@@ -64,7 +67,12 @@ fun UnioApp() {
       composable(Screen.ACCOUNT_DETAILS) { AccountDetails(navigationActions) }
     }
     navigation(startDestination = Screen.HOME, route = Route.HOME) {
-      composable(Screen.HOME) { HomeScreen(navigationActions) }
+      composable(Screen.HOME) {
+        HomeScreen(navigationActions,
+          EventListViewModel(EventRepositoryFirestore(FirebaseFirestore.getInstance())),
+          onAddEvent = {},
+          onEventClick = {})
+      }
     }
     navigation(startDestination = Screen.EXPLORE, route = Route.EXPLORE) {
       composable(Screen.EXPLORE) { ExploreScreen(navigationActions) }
